tools/pyproject.toml
[project]
name = "placeholder-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "George sena",email = "exssgg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
